Ships weaponised
Weapon functions added to Ships.

New ship types include turrets, able to aim and fire munitions; and
various munitions, modeled as ships, which can be proxy-fused or guided,
and kill enemy ships.

Current program demos this via Gunboat ship. This follows the mouse, and
has a missile launcher and a laser turret on the back. Additionally,
press leftMouse to launch a missile from the cursor, or rightMouse to
fire a random laser bolt.