MoreShaderWork: Complete rendering pipeline
Incorporated transform and sprite technology from Ships.

Added light stencil asset.

Most importantly, the system now runs via a black-box renderer. Feed it
sprites and lights, tell it to render, and it spits out a fully
composited lit scene. The program stress tests the process by painting
64 256*256 animated starships across the screen, and putting 256 lights
across them. It gets 20-30fps. Reducing the count or resolution doesn't
have a strictly linear effect, but it's a good benchmark.

Move the mouse to reposition the lights.

This code will be integrated into Ships shortly. The new classes Light
and RenderManager and the asset lightStencil are the novel inclusions.
Note that the Sprite class was enhanced to include discrete diffuse,
specular, emissive, and normal values, and should also be reimported.