ShipsShaders particle updates
Lots of work on the particle systems and assorted bits.

Motion cursor now uses a shader for rapid visualisation.

Particle now doesn't rely on subclasses. It's always a sprite-based
renderer, with built-in alpha fades and dispersion.

Some tweaks to RenderManager to avoid contamination between sprites or
frames. Additionally, added a background warp texture to cause lensing
on screen edges.

ParticleEmitter created. It accepts a list of templates, then dispenses
a random template from its bag of tricks. Very useful.

Ship updated to use ParticleEmitter. Numerous behavioural tweaks.
Incorporated a large list of template particles for both sides to use.

Core program now manages respawning ship population, and draws a
detailed HUD on the screen to clue the player in to what they're
supposed to be doing.

Program is now animal-tested, thanks to Twitch the rat who sat and
stared at the camera. Twitch may be partially blind, which is
unfortunate for a visual medium.