Beta upgrades: particle blends, thrusters
Sprites now have dual tint fields, allowing one to be animated in a
particle and the other to remain constant. This has greatly simplified
the asset library, which is now simply alpha masks that can be tinted
for each faction or purpose.

Ships (and missiles!) now have manoeuvre thrusters with exhaust plumes
that scale depending on turn rate. Looks pretty neat. Not physically
accurate, but this isn't a Newtonian simulation.