Beta upgrades: dreadnaughts
Converted the old "gunboat" design to something more fittingly horrific,
and added it to the enemy spawn list. It makes good use of the high-res
assets.

High-res assets reapplied. The asset-scaling loader had problems:
specifically, it sometimes lost track of which buffer it was drawing to,
and would either crash or load the wrong stuff into buffers. So far as I
can tell, this is a drawback to repeatedly creating temporary buffers in
the setup() method of Processing itself. There is no noteworthy
performance hit from this change, and memory use is halved. I think
that's worth it.

Added some clarification: weapons no longer charge while under cloak.

Cloak pulses slowly. The eagle-eyed might spot it lying in wait.

Shrunk enemy turrets. They look better on the dreadnaughts, and weren't
particularly visible on the smaller vessels.